+++ src/main/java/teammates/ui/webapi/action/BinFeedbackSessionAction.java
@@ -6,6 +6,7 @@ import teammates.common.datatransfer.attributes.FeedbackSessionAttributes;
 import teammates.common.exception.EntityDoesNotExistException;
 import teammates.common.exception.InvalidParametersException;
 import teammates.common.util.Const;
+import teammates.ui.webapi.output.FeedbackSession;
 
 /**
  * Move the feedback sesion to the recycle bin.
@@ -41,7 +42,7 @@ public class BinFeedbackSessionAction extends Action {
         }
 
         FeedbackSessionAttributes recycleBinFs = logic.getFeedbackSessionFromRecycleBin(feedbackSessionName, courseId);
-        return new JsonResult(new FeedbackSessionInfo.FeedbackSessionResponse(recycleBinFs));
+        return new JsonResult(new FeedbackSession(recycleBinFs));
     }
 
 }
