+++ src/main/java/teammates/ui/webapi/action/GetFeedbackResponsesAction.java
@@ -8,6 +8,7 @@ import teammates.common.exception.EntityDoesNotExistException;
 import teammates.common.exception.EntityNotFoundException;
 import teammates.common.exception.InvalidHttpParameterException;
 import teammates.common.util.Const;
+import teammates.ui.webapi.output.FeedbackResponses;
 
 /**
  * Get all responses given by the user for a question.
@@ -55,16 +56,16 @@ public class GetFeedbackResponsesAction extends BasicFeedbackSubmissionAction {
         Intent intent = Intent.valueOf(getNonNullRequestParamValue(Const.ParamsNames.INTENT));
         FeedbackQuestionAttributes questionAttributes = logic.getFeedbackQuestion(feedbackQuestionId);
 
-        FeedbackResponseInfo.FeedbackResponsesResponse result;
+        FeedbackResponses result;
         switch (intent) {
         case STUDENT_SUBMISSION:
             StudentAttributes studentAttributes = getStudentOfCourseFromRequest(questionAttributes.getCourseId());
-            result = new FeedbackResponseInfo.FeedbackResponsesResponse(
+            result = new FeedbackResponses(
                     logic.getFeedbackResponsesFromStudentOrTeamForQuestion(questionAttributes, studentAttributes));
             break;
         case INSTRUCTOR_SUBMISSION:
             InstructorAttributes instructorAttributes = getInstructorOfCourseFromRequest(questionAttributes.getCourseId());
-            result = new FeedbackResponseInfo.FeedbackResponsesResponse(
+            result = new FeedbackResponses(
                     logic.getFeedbackResponsesFromInstructorForQuestion(questionAttributes, instructorAttributes));
             break;
         default:
