+++ src/main/java/teammates/ui/webapi/action/GetFeedbackSessionAction.java
@@ -5,6 +5,7 @@ import teammates.common.datatransfer.attributes.InstructorAttributes;
 import teammates.common.datatransfer.attributes.StudentAttributes;
 import teammates.common.exception.InvalidHttpParameterException;
 import teammates.common.util.Const;
+import teammates.ui.webapi.output.FeedbackSession;
 
 /**
  * Get a feedback session.
@@ -49,8 +50,8 @@ public class GetFeedbackSessionAction extends BasicFeedbackSubmissionAction {
         FeedbackSessionAttributes feedbackSession = logic.getFeedbackSession(feedbackSessionName, courseId);
         Intent intent = Intent.valueOf(getNonNullRequestParamValue(Const.ParamsNames.INTENT));
 
-        FeedbackSessionInfo.FeedbackSessionResponse response =
-                new FeedbackSessionInfo.FeedbackSessionResponse(feedbackSession);
+        FeedbackSession response =
+                new FeedbackSession(feedbackSession);
 
         switch (intent) {
         case STUDENT_SUBMISSION:
