+++ src/main/java/teammates/ui/webapi/action/RestoreFeedbackSessionAction.java
@@ -7,6 +7,7 @@ import teammates.common.exception.EntityDoesNotExistException;
 import teammates.common.exception.EntityNotFoundException;
 import teammates.common.exception.InvalidParametersException;
 import teammates.common.util.Const;
+import teammates.ui.webapi.output.FeedbackSession;
 
 /**
  * Restore a feedback session from the recycle bin.
@@ -48,7 +49,7 @@ public class RestoreFeedbackSessionAction extends Action {
         }
 
         FeedbackSessionAttributes restoredFs = logic.getFeedbackSession(feedbackSessionName, courseId);
-        return new JsonResult(new FeedbackSessionInfo.FeedbackSessionResponse(restoredFs));
+        return new JsonResult(new FeedbackSession(restoredFs));
     }
 
 }
