+++ src/main/java/teammates/ui/webapi/action/SaveFeedbackResponseAction.java
@@ -14,6 +14,7 @@ import teammates.common.exception.InvalidHttpParameterException;
 import teammates.common.exception.InvalidHttpRequestBodyException;
 import teammates.common.exception.UnauthorizedAccessException;
 import teammates.common.util.Const;
+import teammates.ui.webapi.output.FeedbackResponse;
 
 /**
  * Save a feedback response.
@@ -115,7 +116,7 @@ public class SaveFeedbackResponseAction extends BasicFeedbackSubmissionAction {
 
         FeedbackResponseAttributes updatedFeedbackResponse = logic.getFeedbackResponse(
                 feedbackQuestion.getId() + "%" + feedbackResponse.giver + "%" + feedbackResponse.recipient);
-        return new JsonResult(new FeedbackResponseInfo.FeedbackResponseResponse(updatedFeedbackResponse));
+        return new JsonResult(new FeedbackResponse(updatedFeedbackResponse));
     }
 
 }
