+++ src/main/java/teammates/ui/webapi/output/FeedbackSessionPublishStatus.java
@@ -1,0 +1,17 @@
+package teammates.ui.webapi.output;
+
+/**
+ * Represents the publish status of the a feedback session.
+ */
+public enum FeedbackSessionPublishStatus {
+
+    /**
+     * Feedback session is published.
+     */
+    PUBLISHED,
+
+    /**
+     * Feedback session is not published.
+     */
+    NOT_PUBLISHED,
+}
