+++ src/main/java/teammates/ui/webapi/output/FeedbackVisibilityType.java
@@ -1,0 +1,31 @@
+package teammates.ui.webapi.output;
+
+/**
+ * The feedback visibility type.
+ */
+public enum FeedbackVisibilityType {
+    /**
+     * General recipient.
+     */
+    RECIPIENT,
+
+    /**
+     * Giver's team member.
+     */
+    GIVER_TEAM_MEMBERS,
+
+    /**
+     * Recipient's team members.
+     */
+    RECIPIENT_TEAM_MEMBERS,
+
+    /**
+     * Students in the course.
+     */
+    STUDENTS,
+
+    /**
+     * Instructors in the course.
+     */
+    INSTRUCTORS,
+}
