+++ src/main/java/teammates/ui/webapi/output/NumberOfEntitiesToGiveFeedbackToSetting.java
@@ -1,0 +1,16 @@
+package teammates.ui.webapi.output;
+
+/**
+ * The setting of number of entities to giver feedback to.
+ */
+public enum NumberOfEntitiesToGiveFeedbackToSetting {
+    /**
+     * Custom number of entities to give feedback to.
+     */
+    CUSTOM,
+
+    /**
+     * Unlimited number of entities to give feedback to.
+     */
+    UNLIMITED,
+}
