+++ src/main/java/teammates/ui/webapi/output/ResponseVisibleSetting.java
@@ -1,0 +1,21 @@
+package teammates.ui.webapi.output;
+
+/**
+ * The option for response visible setting.
+ */
+public enum ResponseVisibleSetting {
+    /**
+     * Customized response visible time.
+     */
+    CUSTOM,
+
+    /**
+     * Response visible when session is visible.
+     */
+    AT_VISIBLE,
+
+    /**
+     * Response won't be visible automatically.
+     */
+    LATER
+}
