+++ src/main/java/teammates/ui/webapi/output/SessionVisibleSetting.java
@@ -1,0 +1,16 @@
+package teammates.ui.webapi.output;
+
+/**
+ * The option for session visible setting.
+ */
+public enum SessionVisibleSetting {
+    /**
+     * Customized session visible time.
+     */
+    CUSTOM,
+
+    /**
+     * Session visible when open.
+     */
+    AT_OPEN
+}
