+++ src/test/java/teammates/test/cases/webapi/GetFeedbackQuestionsActionTest.java
@@ -11,10 +11,13 @@ import teammates.common.datatransfer.attributes.FeedbackSessionAttributes;
 import teammates.common.datatransfer.attributes.InstructorAttributes;
 import teammates.common.util.Const;
 import teammates.common.util.JsonUtils;
-import teammates.ui.webapi.action.FeedbackQuestionInfo;
 import teammates.ui.webapi.action.GetFeedbackQuestionsAction;
 import teammates.ui.webapi.action.Intent;
 import teammates.ui.webapi.action.JsonResult;
+import teammates.ui.webapi.output.FeedbackQuestion;
+import teammates.ui.webapi.output.FeedbackQuestions;
+import teammates.ui.webapi.output.FeedbackVisibilityType;
+import teammates.ui.webapi.output.NumberOfEntitiesToGiveFeedbackToSetting;
 
 /**
  * SUT: {@link GetFeedbackQuestionsAction}.
@@ -59,13 +62,12 @@ public class GetFeedbackQuestionsActionTest extends BaseActionTest<GetFeedbackQu
         JsonResult r = getJsonResult(a);
 
         assertEquals(HttpStatus.SC_OK, r.getStatusCode());
-        FeedbackQuestionInfo.FeedbackQuestionsResponse feedbackQuestionsResponse =
-                (FeedbackQuestionInfo.FeedbackQuestionsResponse) r.getOutput();
+        FeedbackQuestions feedbackQuestionsResponse = (FeedbackQuestions) r.getOutput();
 
-        List<FeedbackQuestionInfo.FeedbackQuestionResponse> questions = feedbackQuestionsResponse.getQuestions();
+        List<FeedbackQuestion> questions = feedbackQuestionsResponse.getQuestions();
         assertEquals(5, questions.size());
 
-        FeedbackQuestionInfo.FeedbackQuestionResponse typicalResponse = questions.get(0);
+        FeedbackQuestion typicalResponse = questions.get(0);
         FeedbackQuestionAttributes expected =
                 logic.getFeedbackQuestionsForSession(feedbackSessionAttributes.getFeedbackSessionName(),
                         feedbackSessionAttributes.getCourseId()).get(0);
@@ -84,15 +86,15 @@ public class GetFeedbackQuestionsActionTest extends BaseActionTest<GetFeedbackQu
         assertEquals(expected.getGiverType(), typicalResponse.getGiverType());
         assertEquals(expected.getRecipientType(), typicalResponse.getRecipientType());
 
-        assertEquals(FeedbackQuestionInfo.NumberOfEntitiesToGiveFeedbackToSetting.CUSTOM,
+        assertEquals(NumberOfEntitiesToGiveFeedbackToSetting.CUSTOM,
                 typicalResponse.getNumberOfEntitiesToGiveFeedbackToSetting());
         assertEquals(1, typicalResponse.getCustomNumberOfEntitiesToGiveFeedbackTo().intValue());
 
-        assertEquals(Arrays.asList(FeedbackQuestionInfo.FeedbackVisibilityType.INSTRUCTORS),
+        assertEquals(Arrays.asList(FeedbackVisibilityType.INSTRUCTORS),
                 typicalResponse.getShowResponsesTo());
-        assertEquals(Arrays.asList(FeedbackQuestionInfo.FeedbackVisibilityType.INSTRUCTORS),
+        assertEquals(Arrays.asList(FeedbackVisibilityType.INSTRUCTORS),
                 typicalResponse.getShowGiverNameTo());
-        assertEquals(Arrays.asList(FeedbackQuestionInfo.FeedbackVisibilityType.INSTRUCTORS),
+        assertEquals(Arrays.asList(FeedbackVisibilityType.INSTRUCTORS),
                 typicalResponse.getShowRecipientNameTo());
     }
